Data Analysis
Household Data Analysis for the information collected, analysis will help in informing the relocation action plan on important cultural aspects to conform to international standards. Also, this information from the data analysis may lead to the establishment of a cultural heritage management plan for the Mahenge graphite project which is essential for smooth project implementation.